+++ src/engine/Graphics/Raster.ts
@@ -1,7 +1,6 @@
 import { Graphic, GraphicOptions } from './Graphic';
 import { ExcaliburGraphicsContext } from './Context/ExcaliburGraphicsContext';
 import { Color } from '../Color';
-import { ensurePowerOfTwo } from './Context/webgl-util';
 import { Vector } from '../Math/vector';
 import { BoundingBox } from '../Collision/BoundingBox';
 import { watch } from '../Util/Watch';
@@ -94,9 +93,8 @@ export abstract class Raster extends Graphic {
     // get the default canvas width/height as a fallback
     const bitmapWidth = options?.width ?? this._bitmap.width;
     const bitmapHeight = options?.height ?? this._bitmap.height;
-    // Rasters use power of two images as an optimization for webgl
-    this.width = ensurePowerOfTwo(bitmapWidth);
-    this.height = ensurePowerOfTwo(bitmapHeight);
+    this.width = bitmapWidth;
+    this.height = bitmapHeight;
     const maybeCtx = this._bitmap.getContext('2d');
     if (!maybeCtx) {
       /* istanbul ignore next */